Fix Conflicts & Move Around Lib Files
